mngr_imbue_cloud: decompose providers/instance.py (phase 2)
Extract cohesive helper clusters out of the 2k-line instance.py into sibling
modules in the providers/ layer:

- providers/listing.py: pure listing-shaping helpers (derive_host_state_from_raw,
  derive_offline_note_from_raw, map_docker_status_to_host_state). Dropped the
  dead _certified_host_name helper.
- providers/wipe.py: build_pool_host_wipe_script + its constants.
- providers/rebuild.py: the slow-path VPS-vs-slice rebuild builders, converted
  from methods to module-level functions taking explicit deps (no instance.py
  import -> no cycle): build_delegated_vps_provider, build_slice_rebuild_provider,
  _build_delegated_vps_config.

ImbueCloudProvider and its self-bound methods stay in instance.py (2013 -> ~1775
lines). Unit tests for the moved helpers move with them (listing_test.py,
wipe_test.py, rebuild_test.py). Pure refactor, no behavior change.

Also corrected the spec's two pre-implementation inaccuracies (no out-of-plugin
importers; the layers contract needs a dedicated meta-ratchet test, not the
default one) and marked it implemented.

253 unit/integration tests, 56 package ratchets, and 20 meta-ratchets
(ty, ruff, CLI docs, both import-layer contracts) all pass.
