+++ libs/mngr_imbue_cloud/changelog/mngr-ovh-exploration.md
@@ -23,3 +23,5 @@ Added the OVH bare-metal "slices" feature: an alternative to ordering OVH VPSes 
 - Removed the duplicated, on-box slice bake. The slice path no longer ships the monorepo + forever-claude-template to the box and runs `mngr create` there; instead a slice is now provisioned (carved) and baked exactly like an OVH VPS, from the operator's machine. `LimaSliceVpsClient` drives `limactl` over SSH on the box (carve a bare Debian VM = the "OS reinstall" equivalent), and the shared container bake then reaches the VM's box-forwarded ports -- so `allocate-slice` just vendors mngr into the FCT workspace once and runs the bake from here. The FCT bake itself (templates, `system-services` agent, chat-agent teardown, sentinel) now lives in one provider-generic place (`pool_bake.py`) shared by both the OVH and slice paths, instead of being copy-pasted into the slice tooling. This deletes `slice_bake.py` and the box-side rsync/`uv sync`/git-init machinery; behavior for the operator is unchanged (`allocate-slice` still parallel-bakes `--count N` slices onto one ready box and inserts their rows).
 
 - Restructured the now-large `mngr_imbue_cloud` plugin from a flat module list into layered sub-packages (`plugin`, `cli`, `bake`, `providers`, `hosts`, `slices`, `connector`) plus the shared root leaf modules (`config`, `data_types`, `errors`, `primitives`), with an `import-linter` "mngr_imbue_cloud layers contract" enforcing the ordering (and a meta-ratchet test that gates it). The slice/bare-metal subsystem is isolated in `slices/`, the provider-generic pool bake in `bake/` (an extraction seam toward the minds app), and both provider backends are co-located in `plugin/backends.py`. Pure refactor: no behavior, CLI, wire-format, or schema change. Plugin entry points moved to `imbue.mngr_imbue_cloud.plugin.entrypoints` / `plugin.slice_entrypoints`.
+
+- Decomposed the oversized `providers/instance.py` (~2,000 lines): extracted the pure listing-shaping helpers into `providers/listing.py`, the pre-release data-wipe script generator into `providers/wipe.py`, and the slow-path VPS-vs-slice rebuild provider/config builders into `providers/rebuild.py` (with their unit tests co-located). `ImbueCloudProvider` and its self-bound methods stay in `instance.py`. Removed a dead helper (`_certified_host_name`). No behavior change.
