+++ libs/mngr_imbue_cloud/imbue/mngr_imbue_cloud/providers/instance.py
@@ -48,7 +48,6 @@ from pydantic import PrivateAttr
 from pydantic import SecretStr
 
 from imbue.concurrency_group.concurrency_group import ConcurrencyGroup
-from imbue.imbue_common.pure import pure
 from imbue.mngr.errors import HostAuthenticationError
 from imbue.mngr.errors import HostNotFoundError
 from imbue.mngr.errors import MngrError
@@ -85,7 +84,6 @@ from imbue.mngr.primitives import HostId
 from imbue.mngr.primitives import HostName
 from imbue.mngr.primitives import HostState
 from imbue.mngr.primitives import ImageReference
-from imbue.mngr.primitives import ProviderBackendName
 from imbue.mngr.primitives import SSHInfo
 from imbue.mngr.primitives import SnapshotId
 from imbue.mngr.primitives import SnapshotName
@@ -113,172 +111,16 @@ from imbue.mngr_imbue_cloud.errors import ImbueCloudLeaseUnavailableError
 from imbue.mngr_imbue_cloud.hosts.host import ImbueCloudHost
 from imbue.mngr_imbue_cloud.primitives import FastMode
 from imbue.mngr_imbue_cloud.primitives import ImbueCloudAccount
-from imbue.mngr_imbue_cloud.providers.slice_provider import SliceVpsDockerProvider
-from imbue.mngr_imbue_cloud.providers.slice_provider import SliceVpsDockerProviderConfig
-from imbue.mngr_imbue_cloud.slices.lima_slice_client import LimaSliceVpsClient
-from imbue.mngr_vps_docker.config import VpsDockerProviderConfig
+from imbue.mngr_imbue_cloud.providers.listing import derive_host_state_from_raw
+from imbue.mngr_imbue_cloud.providers.listing import derive_offline_note_from_raw
+from imbue.mngr_imbue_cloud.providers.rebuild import build_delegated_vps_provider
+from imbue.mngr_imbue_cloud.providers.rebuild import build_slice_rebuild_provider
+from imbue.mngr_imbue_cloud.providers.wipe import build_pool_host_wipe_script
 from imbue.mngr_vps_docker.host_setup import apply_host_setup_on_outer
-from imbue.mngr_vps_docker.instance import MinimalVpsDockerProvider
 from imbue.mngr_vps_docker.instance import VpsDockerProvider
 from imbue.mngr_vps_docker.primitives import VpsInstanceId
-from imbue.mngr_vps_docker.vps_client import ExternallyManagedVpsClient
 
 _SSH_WAIT_TIMEOUT_SECONDS: Final[float] = 120.0
-
-# Path on every mngr_vps_docker-baked outer where the per-host btrfs loop
-# filesystem is mounted (matches ``VpsDockerProviderConfig.btrfs_mount_path``
-# default). The per-host subvolume is at ``<this>/<host_id_hex>``.
-_VPS_BTRFS_MOUNT_PATH: Final[str] = "/mngr-btrfs"
-
-# Container label that mngr_vps_docker tags the workspace container with;
-# also used as ``docker volume`` -name search prefix (the bind-options volume
-# is named ``mngr-host-vol-<host_id_hex>``).
-_LABEL_HOST_ID_KEY: Final[str] = "com.imbue.mngr.host-id"
-_HOST_VOLUME_NAME_PREFIX: Final[str] = "mngr-host-vol-"
-
-
-def build_pool_host_wipe_script(host_id: HostId) -> str:
-    """Render the bash script that wipes a leased pool VPS before release.
-
-    Runs as root on the outer (the leased VPS itself). Each step is
-    independently best-effort -- the script always exits 0 so a single
-    failed sub-step (e.g. ``docker stop`` on an already-stopped container)
-    doesn't abort the rest. The destroy-host caller still logs warnings
-    when this returns abnormally, but the privacy-relevant steps remain
-    unconditionally attempted.
-
-    Wipes:
-
-    * The workspace container (by label) and any named docker volume
-      whose name contains the host hex.
-    * The per-host btrfs subvolume backing the bind-options volume (the
-      bind source ``device=`` lives outside the container, so removing
-      the container alone leaves the data on disk).
-    * Everything docker still knows about (``system prune -a -f --volumes``).
-    * Everything under ``/root`` and ``/tmp`` except
-      ``/root/.ssh/authorized_keys`` (preserves the static pool-management
-      key the connector / cleanup_released_hosts.py needs for any
-      post-release SSH-driven maintenance).
-
-    Pure function so unit tests can assert the exact command shape
-    without standing up an SSH transport.
-    """
-    host_id_hex = host_id.get_uuid().hex
-    host_id_str = str(host_id)
-    label_filter = shlex.quote(f"label={_LABEL_HOST_ID_KEY}={host_id_str}")
-    volume_name_filter = shlex.quote(f"name={_HOST_VOLUME_NAME_PREFIX}{host_id_hex}")
-    subvolume_path = f"{_VPS_BTRFS_MOUNT_PATH}/{host_id_hex}"
-    return (
-        "set +e\n"
-        # Find + remove the workspace container by its host-id label.
-        f"container_id=$(docker ps -a --filter {label_filter} --format '{{{{.ID}}}}' | head -1)\n"
-        'if [ -n "$container_id" ]; then\n'
-        '    docker stop "$container_id" >/dev/null 2>&1\n'
-        '    docker rm -f -v "$container_id" >/dev/null 2>&1\n'
-        "fi\n"
-        # Drop the per-host named volume(s). docker volume rm -f tolerates
-        # the volume not existing.
-        f"for vol in $(docker volume ls -q --filter {volume_name_filter}); do\n"
-        '    docker volume rm -f "$vol" >/dev/null 2>&1\n'
-        "done\n"
-        # The bind-options volume's storage is a btrfs subvolume at
-        # /mngr-btrfs/<host_hex>; docker volume rm doesn't touch the bind
-        # source, so wipe it explicitly. Fall back to rm -rf if btrfs is
-        # absent (older non-btrfs hosts before the btrfs spec landed).
-        f"if [ -d {shlex.quote(subvolume_path)} ]; then\n"
-        f"    btrfs subvolume delete {shlex.quote(subvolume_path)} >/dev/null 2>&1 || "
-        f"rm -rf {shlex.quote(subvolume_path)} >/dev/null 2>&1\n"
-        "fi\n"
-        # Reclaim everything else docker holds: stopped containers, dangling
-        # images, networks, unused volumes, build cache.
-        "docker system prune -a -f --volumes >/dev/null 2>&1\n"
-        # Wipe /root content except .ssh/authorized_keys -- preserves the
-        # pool-management public key the connector relies on if it needs
-        # to reach the host again before cleanup_released_hosts.py runs.
-        "find /root -mindepth 1 -maxdepth 1 -not -name .ssh -exec rm -rf {} + 2>/dev/null\n"
-        "find /root/.ssh -mindepth 1 -not -name authorized_keys -exec rm -rf {} + 2>/dev/null\n"
-        # Wipe /tmp entirely.
-        "find /tmp -mindepth 1 -maxdepth 1 -exec rm -rf {} + 2>/dev/null\n"
-        # Always succeed: release is the gating step, individual wipe
-        # failures are surfaced via the script's stderr but must not
-        # block lease return.
-        "exit 0\n"
-    )
-
-
-def _certified_host_name(raw: Mapping[str, Any]) -> str | None:
-    """Pull the friendly host name out of the listing raw output, if present."""
-    certified = raw.get("certified_data") or {}
-    name = certified.get("host_name")
-    return name if isinstance(name, str) and name else None
-
-
-def _derive_host_state_from_raw(raw: Mapping[str, Any]) -> HostState:
-    """Map the outer-listing raw output to a HostState.
-
-    The outer listing script tags the output with ``CONTAINER_STATE``,
-    ``CONTAINER_EXIT_CODE``, and ``CONTAINER_MISSING`` so we don't have
-    to re-run docker inspect.
-    """
-    if raw.get("container_missing"):
-        return HostState.DESTROYED
-    container_state = raw.get("container_state")
-    if not container_state:
-        # Outer SSH succeeded but produced no state -- treat as crashed
-        # (no info to be more specific).
-        return HostState.CRASHED
-    exit_code = raw.get("container_exit_code") or 0
-    has_certified_data = bool(raw.get("certified_data"))
-    if container_state == "running" and has_certified_data:
-        return HostState.RUNNING
-    if container_state == "running":
-        # Container is up but docker exec didn't give us data -- we know
-        # the host exists but can't read its state from inside.
-        return HostState.UNAUTHENTICATED
-    state, _note = _map_docker_status_to_host_state(container_state, exit_code)
-    return state
-
-
-def _derive_offline_note_from_raw(raw: Mapping[str, Any]) -> str | None:
-    """Produce a short ``failure_reason`` note for non-running containers.
-
-    Returns None for running containers (no note needed) and for the
-    DESTROYED / missing case (the state itself is the message). For
-    stopped/paused/etc., returns the human-readable note that
-    ``_map_docker_status_to_host_state`` produced.
-    """
-    container_state = raw.get("container_state")
-    if not container_state or container_state == "running":
-        return None
-    if raw.get("container_missing"):
-        return None
-    exit_code = raw.get("container_exit_code") or 0
-    _state, note = _map_docker_status_to_host_state(container_state, exit_code)
-    return note
-
-
-def _map_docker_status_to_host_state(status: str, exit_code: int) -> tuple[HostState, str | None]:
-    """Translate docker's container ``State.Status`` into a ``HostState``.
-
-    Returns ``(state, note)`` where ``note`` is a short human-readable
-    diagnostic appended to ``HostDetails.failure_reason``. If the docker
-    container is ``running`` but inner SSH was unreachable we treat that
-    as an authentication problem -- the host is up; we just can't get
-    inside it.
-    """
-    if status == "running":
-        return HostState.UNAUTHENTICATED, "container is running on outer host but inner SSH was unreachable"
-    if status == "exited":
-        if exit_code == 0:
-            return HostState.STOPPED, "container exited cleanly"
-        return HostState.CRASHED, f"container exited with code {exit_code}"
-    if status == "paused":
-        return HostState.PAUSED, "container is paused"
-    if status in ("created", "restarting"):
-        return HostState.STARTING, f"container in {status} state"
-    if status in ("dead", "removing"):
-        return HostState.CRASHED, f"container in {status} state"
-    return HostState.CRASHED, f"unrecognized docker status {status!r}"
 
 
 def _rewrite_container_host_name(
@@ -383,24 +225,6 @@ def _scan_ssh_host_key(host: str, port: int) -> str | None:
         except (OSError, paramiko.SSHException):
             pass
     return f"{host_key.get_name()} {host_key.get_base64()}"
-
-
-@pure
-def _build_delegated_vps_config(config: ImbueCloudProviderConfig) -> VpsDockerProviderConfig:
-    """Build the delegated vps_docker config for the slow-path rebuild.
-
-    Forwards the runtime knobs (``docker_runtime`` / ``install_gvisor_runtime`` /
-    ``default_start_args``) from the imbue_cloud config so the rebuilt container
-    runs under the configured runtime with the configured hardening args.
-    """
-    return VpsDockerProviderConfig(
-        backend=ProviderBackendName("vps_docker"),
-        host_dir=config.host_dir,
-        container_ssh_port=config.container_ssh_port,
-        docker_runtime=config.docker_runtime,
-        install_gvisor_runtime=config.install_gvisor_runtime,
-        default_start_args=config.default_start_args,
-    )
 
 
 class ImbueCloudProvider(BaseProviderInstance):
@@ -629,7 +453,7 @@ class ImbueCloudProvider(BaseProviderInstance):
                 result[host_ref] = agent_refs
                 continue
             self._listing_raw_cache[host_id] = raw
-            host_state = _derive_host_state_from_raw(raw)
+            host_state = derive_host_state_from_raw(raw)
             if host_state == HostState.DESTROYED and not include_destroyed:
                 continue
             # ``entry.host_name`` is the canonical user-supplied name from the
@@ -887,8 +711,8 @@ class ImbueCloudProvider(BaseProviderInstance):
         by ``build_outer_listing_collection_script``).
         """
         ssh_info = self._build_lease_ssh_info(host_ref.host_id, lease)
-        host_state = _derive_host_state_from_raw(raw)
-        failure_reason = _derive_offline_note_from_raw(raw)
+        host_state = derive_host_state_from_raw(raw)
+        failure_reason = derive_offline_note_from_raw(raw)
         boot_time = timestamp_to_datetime(raw.get("btime"))
         uptime_seconds = raw.get("uptime_seconds")
         lock_mtime = raw.get("lock_mtime")
@@ -1380,7 +1204,11 @@ class ImbueCloudProvider(BaseProviderInstance):
         # publish port -- true only for slices (forwarded ports), never OVH VPSes.
         is_slice = lease_result.container_ssh_port != self.config.container_ssh_port
         delegated_provider: VpsDockerProvider = (
-            self._build_slice_rebuild_provider(lease_result) if is_slice else self._build_delegated_vps_provider()
+            build_slice_rebuild_provider(
+                name=self.name, config=self.config, mngr_ctx=self.mngr_ctx, lease_result=lease_result
+            )
+            if is_slice
+            else build_delegated_vps_provider(name=self.name, config=self.config, mngr_ctx=self.mngr_ctx)
         )
         # The VPS root host key (port 22 for OVH; the box-forwarded VM-root port
         # for a slice) feeds the rebuilt host's record AND is pinned in the
@@ -1433,73 +1261,6 @@ class ImbueCloudProvider(BaseProviderInstance):
                 known_hosts=known_hosts,
                 authorized_keys=combined_authorized_keys,
             )
-
-    def _build_delegated_vps_provider(self) -> VpsDockerProvider:
-        """Construct a vps_docker provider bound to this instance's keys/config.
-
-        It only ever runs ``teardown_container_on_existing_vps`` /
-        ``create_host_on_existing_vps`` (which take a caller-supplied ``outer``
-        and make no VPS-API calls), so its ``vps_client`` is the
-        ``ExternallyManagedVpsClient`` stub that raises on any ordering call.
-
-        Forwards the runtime knobs from ``self.config`` (an
-        ``ImbueCloudProviderConfig``, which extends ``VpsDockerProviderConfig``)
-        so the rebuilt container runs under the configured runtime with the
-        configured hardening args -- e.g. ``docker_runtime='runsc'`` plus
-        ``--workdir=/`` / ``--security-opt=no-new-privileges`` from
-        ``default_start_args``, which minds bootstrap writes into the per-account
-        block.
-        """
-        vps_config = _build_delegated_vps_config(self.config)
-        return MinimalVpsDockerProvider(
-            name=self.name,
-            host_dir=self.config.host_dir,
-            mngr_ctx=self.mngr_ctx,
-            config=vps_config,
-            vps_client=ExternallyManagedVpsClient(),
-        )
-
-    def _build_slice_rebuild_provider(self, lease_result: LeaseResult) -> SliceVpsDockerProvider:
-        """Construct a slice provider to rebuild the container on a leased slice VM.
-
-        A slice's container is published inside the VM on the standard guest port
-        (``container_ssh_port``, which lima forwards to a box host port) but is
-        reached from outside at the lease's forwarded ``container_ssh_port`` / VM
-        root ``ssh_port``. The slice provider already splits publish vs connect
-        ports via these per-host-port fields, so the rebuild (teardown +
-        ``create_host_on_existing_vps``) targets the right ports. runsc/gVisor is
-        not used (the VM is the isolation boundary; its Docker is plain runc).
-        """
-        slice_config = SliceVpsDockerProviderConfig(
-            host_dir=self.config.host_dir,
-            container_ssh_port=self.config.container_ssh_port,
-            box_public_address=lease_result.vps_address,
-        )
-        # The rebuild never carves/destroys a VM (it only tears down + rebuilds the
-        # container on the already-leased slice via the forwarded ports below), so
-        # the lima client's box-SSH coordinates are unused here; pass the address
-        # for completeness and no pool key (limactl is never invoked on this path).
-        lima_client = LimaSliceVpsClient(
-            box_address=lease_result.vps_address,
-            box_ssh_user=slice_config.box_ssh_user,
-            private_key_path=None,
-        )
-        provider = SliceVpsDockerProvider(
-            name=self.name,
-            host_dir=self.config.host_dir,
-            mngr_ctx=self.mngr_ctx,
-            config=slice_config,
-            vps_client=lima_client,
-            slice_config=slice_config,
-            lima_client=lima_client,
-        )
-        # Point the per-host-port seams at the lease's box-forwarded ports so the
-        # rebuild's outer (VM root) and container connections target the box.
-        provider.set_forwarded_ports(
-            outer_port=lease_result.ssh_port,
-            container_port=lease_result.container_ssh_port,
-        )
-        return provider
 
     @contextmanager
     def _outer_for_leased_vps(self, host_id: HostId, lease_result: LeaseResult) -> Iterator[OuterHostInterface]:
